add timestamp log
