+++ blog-backend/blog-repository/src/main/java/com/github/iofu728/blog/repository/util/UtilsHelper.java
@@ -1,0 +1,23 @@
+package com.github.iofu728.blog.repository.util;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+public class UtilsHelper {
+
+    public static boolean isNumeric(String s) {
+        for (int i = 0; i < s.length(); i++) {
+            if (!Character.isDigit(s.charAt(i))) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public static String getCurrentTimeStamp() {
+        SimpleDateFormat sdfDate = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+        Date now = new Date();
+        String strDate = sdfDate.format(now);
+        return strDate;
+    }
+}
